requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.2
django-braces==1.16.0
django-crispy-forms==2.3
django-simple-history==3.7.0
mysqlclient==2.2.4
sqlparse==0.5.1

